+++ example/bin/main.dart
@@ -1,0 +1,3 @@
+import 'package:learmond_example/main.dart' as example;
+
+void main() => example.runExample();