+++ CHANGELOG.md
@@ -1,7 +1,40 @@
+## The Learmond Corporation's CLI tool
+
+## 0.0.0
+
+- Pre-release
+
 ## 0.0.1
 
-- The Learmond Corporation's CLI tool. 
+- The Learmond Corporation's official CLI tool. 
 - Initial version.
+
+## 0.0.1+1 
+
+- Stable release.
+
+## 0.0.1+2
+
+- Stable release.
+
+## 0.0.1+3
+
+- Stable release.
+
+## 0.0.1+4
+
+- Stable release.
+
+## 0.1.0
+
+- Breaking release. Fully supports npx/npm, flutter, go and ruby apps. Run learmond <command> <app-name> --<type> to create, test, analyze/lint and run your. Git commands include learmond repo init to initialize a Git repository in your current direct and learmond push to Stage, Commit, and push changes to your newly created branch. 
+
+
+
+
+
+
+